Move flag handling to its own module
